remove instances before deploy
Deploying the next version before changing the alias would result in too many running instances, which exceed the open-source plan of now.sh. For this reason only, we need to remove the active deployment before deploying the new version. This results in some downtime when deploying. For a real production app, it would be done like this: deploy first, then change alias, then optionally remove the old instace (or keep it for a rollback).